pyproject.toml
[tool.poetry]
name = "educationapp"
version = "0.1.0"
description = ""
authors = ["Ekaterina Itchenko <ekaterina.savickaya.1989@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.10"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
